Adds install recipe and alias
Adds a cargo install recipe to the Justfile. This allows
the user to easily install the application.

It also adds an alias for the install recipe to improve
usability.

Sets the maximum tracing level to TRACE. This enables
more verbose logging for debugging purposes. The
environment filter is temporarily disabled.